+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 2.0.0.dev4 (unreleased)
 -----------------------
 
+- Add @booking-move restapi
+  [mamico]
+
 - Extend @@bookings search view parameters list.
   [folix-01]
 - Added event handler on booking creation to send email to managers.